website/website-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile website/website-requirements.in
#
cachetools==4.2.2
    # via google-auth
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
google-api-core[grpc]==1.26.3
    # via google-cloud-pubsub
google-auth==1.30.0
    # via google-api-core
google-cloud-pubsub==2.4.1
    # via -r website/website-requirements.in
googleapis-common-protos[grpc]==1.53.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
grpc-google-iam-v1==0.12.3
    # via google-cloud-pubsub
grpcio==1.37.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
idna==2.10
    # via requests
libcst==0.3.18
    # via google-cloud-pubsub
mypy-extensions==0.4.3
    # via typing-inspect
packaging==20.9
    # via google-api-core
proto-plus==1.18.1
    # via google-cloud-pubsub
protobuf==3.15.8
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyparsing==2.4.7
    # via packaging
pytz==2021.1
    # via google-api-core
pyyaml==5.4.1
    # via libcst
requests==2.25.1
    # via google-api-core
rsa==4.7.2
    # via google-auth
six==1.15.0
    # via
    #   google-api-core
    #   google-auth
    #   grpcio
    #   protobuf
typing-extensions==3.7.4.3
    # via
    #   libcst
    #   typing-inspect
typing-inspect==0.6.0
    # via libcst
urllib3==1.26.4
    # via requests
whitenoise==5.2.0
    # via -r website/website-requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
